Adding feature to be a pull request, test work well, postman and Dynamo works
