+++ server/src/controller/CommentsHelper.js
@@ -20,6 +20,24 @@ function validateNewComment(submittedComment) {
     }
 }
 
+// Validation Middleware
+function validationMiddleware(req, res, next) {
+    try {
+        if(req.path === '/') {
+            if(req.method === 'POST') {
+                req.body.username = req.user.username;
+                //validate the data from the req.body is valid
+                req.body = validateNewComment(req.body);
+            }
+        }
+
+        next();
+    } catch(err) {
+        res.json({message: err.message});
+    }
+}
+
 module.exports = {
-    validateNewComment
+    validateNewComment,
+    validationMiddleware
 }
