+++ server/src/controller/CommentsRouter.js
@@ -1,12 +1,13 @@
 const express = require('express');
 const router = express.Router();
 const commentsService = require('../service/CommentsService');
-const { validateNewComment } = require('./CommentsHelper')
+const { validationMiddleware } = require('./CommentsHelper')
 const { authenticateToken } = require('../util/WebToken');
 
 // CREATE
 // New Comment Post
-router.post('/', validationMiddleware, async (req, res) => {
+//will add authorization check
+router.post('/', authenticateToken, validationMiddleware, async (req, res) => {
     //either throws an error or returns the comment it stores
     try {
         //add username to body through the jwt tokens username
@@ -28,23 +29,5 @@ router.post('/', validationMiddleware, async (req, res) => {
 // UPDATE
 
 // DELETE
-
-
-// Validation Middleware
-function validationMiddleware(req, res, next) {
-    try {
-        if(req.path === '/') {
-            if(req.method === 'POST') {
-                //req.body.username = req.user.username;
-                //validate the data from the req.body is valid
-                req.body = validateNewComment(req.body);
-            }
-        }
-
-        next();
-    } catch(err) {
-        res.json({message: err.message});
-    }
-}
 
 module.exports = router;
